+++ Services/DB/QuoteService.cs
@@ -28,7 +28,11 @@ namespace prod_server.Services.DB
 
         public Task<List<Quote>> GetQuotes()
         {
-            return _database.Quotes.OrderByDescending(q => q.CreatedAt).Include(x => x.Name).ToListAsync();
+            return _database.Quotes
+                    .OrderByDescending(q => q.CreatedAt)
+                    .Where(x => x.isArchived == false)
+                    .Include(x => x.Name)
+                    .ToListAsync();
         }
 
         public Task<List<Quote>> GetWithProducts()
@@ -40,6 +44,7 @@ namespace prod_server.Services.DB
         private IQueryable<Quote> GetQuotesWithProducts()
         {
             return _database.Quotes
+                .Where(x => x.isArchived == false)
                 .Include(q => q.Name)
                 .Include(q => q.Products)
                 .ThenInclude(cp => cp.Product);
@@ -55,6 +60,20 @@ namespace prod_server.Services.DB
         {
             var quote = _database.Quotes.FirstOrDefault(quote => quote.Id == id);
             if (quote == null) throw new Exception("Quote not found.");
+
+            // Check if customer has any relationships with other entities
+            var relatedEntities = _database.Entry(quote)
+                .Metadata
+                .GetNavigations()
+                .Where(n => !n.IsCollection)
+                .Select(n => _database.Entry(quote)?.Reference(n.Name)?.TargetEntry?.Entity)
+                .ToList();
+
+            if (relatedEntities.Any())
+            {
+                // Customer has related entities, handle accordingly (throw exception, log, etc.)
+                throw new InvalidOperationException("Quote cannot be deleted because it has related entities.");
+            }
 
             _database.Remove(quote);
             await _database.SaveChangesAsync();
